dashboard_query.sql - montreal_neuro_ver1e_2023_05_26 #COK-396
Made some SQL changes to make the query more robust for nested fields and which was causing extra rows in the GoogleSheet, but not incorrect data on the dashboard. See attached comments that I made on Slack.

This impacted on the “container_title” for when we had chapters or conference proceedings. 

fixed duplicated  ‘crossref_trn.clinical_trial_number’ and ‘crossref_trn.registry’ fields,